backend/requirements.txt
flask
flask-cors
tensorflow-cpu
pillow
numpy
pytest
requests
gunicorn
python-dotenv
google-generativeai